requirements.txt
bcrypt==5.0.0
blinker==1.8.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.11
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.4
SQLAlchemy==2.0.44
typing_extensions==4.13.2
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2
